Added Operators, UnitTest code and fixed memory leak issues

+++ MonsterChase/Engine/Point2D.cpp
@@ -8,7 +8,7 @@ Point2D::Point2D() {
 }
 
 //Overloaded constructor for direct input
-Point2D::Point2D(int x ,int y)
+Point2D::Point2D(float x ,float y)
 {
 	this->x = x;
 	this->y = y;
